# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/recepie": {
            "get": {
                "description": "get all Recepies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie"
                ],
                "summary": "Get all Recepies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Recepie"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieSID": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie"
                ],
                "summary": "Create an Recepie",
                "parameters": [
                    {
                        "description": "Recepie name",
                        "name": "recepie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.recepieCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recepie"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/recepie/{id}": {
            "get": {
                "description": "get by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie"
                ],
                "summary": "Show an Recepie",
                "operationId": "get-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recepie"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieSID": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie"
                ],
                "summary": "Update an Recepie",
                "parameters": [
                    {
                        "description": "Recepie name",
                        "name": "recepie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.recepieEdit"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recepie"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "boject"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "boject"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieSID": []
                    }
                ],
                "description": "delete by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recepie"
                ],
                "summary": "Delete an Recepie",
                "operationId": "delete-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recepie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "boject"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "boject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.recepieCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "time_to_cook": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "handlers.recepieEdit": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "time_to_cook": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "models.Recepie": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "time_to_cook": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieSID": {
            "type": "apiKey",
            "name": "sid",
            "in": "cookie"
        }
    }
}